Moved wordcloud series to subfolder.
